requirements.txt
absl-py==1.1.0
aiosignal==1.2.0
attrs==21.4.0
certifi==2023.07.22
charset-normalizer==2.0.12
click==8.0.4
distlib==0.3.4
filelock==3.6.0
frozenlist==1.4.0
greenlet==1.1.3
grpcio==1.53.0
idna==3.3
importlib-resources==5.8.0
jsonschema==4.6.0
msgpack==1.0.4
numpy==1.26.4
ortools
pandas==2.2.1
platformdirs==2.6.2
protobuf
psycopg2==2.9.9
pyarrow==15.0.2
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.1
ray==2.8.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.28
urllib3==1.26.18
virtualenv
zipp==3.17.0

pip==24.0
faiss-cpu==1.7.4
boto3==1.28.27
jq==1.4.1
tiktoken==0.4.0
yake==0.4.8
python-dotenv==1.0.0
pandas==2.0.3
google-cloud-aiplatform==1.42.0
semantic-kernel==0.5.1.dev0
openai==1.12.0
# ray[default]==2.9.3

scikit-learn
tokenizers
transformers
bitsandbytes
sentencepiece==0.1.99
cffi
lxml
bs4
unstructured==0.6.9
tabulate
pdf2image
accelerate>=0.12.0
langchain==0.1.7
langchain-community==0.0.28
tiktoken
chromadb
requests
numpy
faiss-cpu
sqlalchemy